Make self-emission model training robust
Improve robustness of self-emission model to outliers.  That means either
outliers on the offset, OR outliers in the temperatures.  This is
currently applied to the training and testing, but still needs to be
applied to the actual evaluation; but instead of having outliers for the
full 24 hour window, this is now limited to outliers for those instances
where the temperatures are actually bad.