Handle missing scanlines and channels
In the CURUC recipes, handle missing scanlines and channels.  This commit
implements an algorithm that first removes channels that are completely
bad, and then removes any scanlines where any of the remaining
otherwise-good channels are bad for that scanline.  It keeps track of the
scanline numbers in order to provide a consistent answer for the
correlations.