Fix bug in propagation beyond summation
Fixed a bug in propagation beyond summation; all_effects is now a
ExpressionDict rather than an ordinary dict, so we can check for an
IndexedBase directly rather than the underlying Symbol.  This is necessary
because I have assigned the effect to the IndexedBase in the previous
commit so that bookkeeping was correct.

It appears that the uncertainty in O_TPRT is now correctly propagated!
Fixes #129.  Closes #129.  Huray!