Re-enable plotting naive radiances
With typhons new feature to customite what exceptions shall be handled,
enable skipping of orbits with duplicates by adding InvalidIndexError from
pandas to this list.  This is needed as a workaround because duplicate
handling is only happening after all granules have been read, but the
naive radiances are calculated after each granule, and this calculation
fails if there are duplicate times due to xarray/pandas time indexing
issues.