When passing covariant components, also pass sensitivity coefficients
When the uncertainty calculation calculated uncertainty components, we
should also return the components due to covariances along with their
sensitivities.  This is needed to later calculate the contribution in
radiance or BT units from the covariances, which is important for analysis
purposes.  We also need the components to subtract from u_structured to
which they have been added initially, in order to put them in u_common
instead.