More flexible harmonisation parameter updates
The script to update the harmonisation parameters now takes a list of
channels and a list of files containing the parameters for those channels.
Previously it was hardwired to be 11 or 19 channels, plus it was quite
risky to get the values wrong if the harmonisation files were not passed
in the order of the channels!